Split electricity preprocessing + correction slow (CE measure) implementation (#553)
* split elc preprocessing into generation, grid, storage

* deleted outdated file

* added docstring to function; deleted outdated functions

* Corrected error in CE measures (slow); moved elc CE measures to separate file in preprocessing/

* minor change, adjusted scenario

* minor change, adjusted scenario

* added docstring to generation preprocessing function
