+++ imagematerials/electricity/preprocessing/__init__.py
@@ -1,0 +1,1 @@
+"""Subpackage containing the preprocessing steps for the elecctricity module."""
